requirements.txt
openai
python-dotenv
pytest
selenium
webdriver-manager